# apiDoc.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "绝缘子裂痕检测API",
    "version": "1.0.0",
    "description": "绝缘子裂痕检测系统的API文档"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发服务器"
    }
  ],
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "用户登录",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {"type": "string"},
                  "password": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {"type": "string"},
                    "user": {"type": "object"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/detect": {
      "post": {
        "summary": "绝缘子裂痕检测",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "绝缘子图片文件"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "检测成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {"type": "boolean"},
                    "cracks": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "x": {"type": "number"},
                          "y": {"type": "number"},
                          "width": {"type": "number"},
                          "height": {"type": "number"},
                          "confidence": {"type": "number"}
                        }
                      }
                    },
                    "suggestions": {
                      "type": "array",
                      "items": {"type": "string"}
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}